Refactored VirtualFile to stand-alone file and added tests
