+++ tests/test_lib.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+"""Functions and classes for testing."""
+
+from __future__ import unicode_literals
+
+import os
+import sys
+import unittest
+
+
+class BaseTestCase(unittest.TestCase):
+  """The base test case."""
+
+  _TEST_DATA_PATH = os.path.join(os.getcwd(), 'test_data')
+
+  # Show full diff results, part of TestCase so does not follow our naming
+  # conventions.
+  maxDiff = None
+
+  def _GetTestFilePath(self, path_segments):
+    """Retrieves the path of a test file relative to the test data directory.
+
+    Args:
+      path_segments (list[str]): path segments inside the test data directory.
+
+    Returns:
+      str: a path of the test file.
+    """
+    # Note that we need to pass the individual path segments to os.path.join
+    # and not a list.
+    return os.path.join(self._TEST_DATA_PATH, *path_segments)
